refactor: separate option texts
